refactor(lyric): 调整歌词播放位置数据类型为Long并重构接口结构
- 将onSeekTo和onPositionChanged方法的position参数从Int改为Long
- 将歌词模型中的时间相关字段从Int改为Long类型
- 重构歌词相关接口分离到独立文件中
- 添加DeepCopyable和Normalize接口支持歌词深拷贝和标准化
- 实现歌词模型的深拷贝和规范化功能
- 添加REQUEST_HIGHLIGHT_VIEW常量
- 添加插值器依赖和工具类
- 优化歌词视图的动画和滚动逻辑
- 移除过时的Syllable实现代码

+++ lyric/bridge/central/src/main/kotlin/io/github/proify/lyricon/central/provider/player/PlayerListener.kt
@@ -21,7 +21,7 @@ import io.github.proify.lyricon.lyric.model.Song
 interface PlayerListener {
     fun onSongChanged(recorder: PlayerRecorder, song: Song?)
     fun onPlaybackStateChanged(recorder: PlayerRecorder, isPlaying: Boolean)
-    fun onPositionChanged(recorder: PlayerRecorder, position: Int)
-    fun onSeekTo(recorder: PlayerRecorder, position: Int)
+    fun onPositionChanged(recorder: PlayerRecorder, position: Long)
+    fun onSeekTo(recorder: PlayerRecorder, position: Long)
     fun onPostText(recorder: PlayerRecorder, text: String?)
 }
