+++ lyric/model/src/main/kotlin/io/github/proify/lyricon/lyric/model/LyricTiming.kt
@@ -17,6 +17,7 @@
 package io.github.proify.lyricon.lyric.model
 
 import android.os.Parcelable
+import io.github.proify.lyricon.lyric.model.interfaces.ILyricTiming
 import kotlinx.parcelize.Parcelize
 import kotlinx.serialization.Serializable
 
@@ -30,7 +31,7 @@ import kotlinx.serialization.Serializable
 @Serializable
 @Parcelize
 data class LyricTiming(
-    override var begin: Int,
-    override var end: Int,
-    override var duration: Int
+    override var begin: Long,
+    override var end: Long,
+    override var duration: Long
 ) : ILyricTiming, Parcelable
